% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MVNSeq.r
\name{hmm2EMStep}
\alias{hmm2EMStep}
\title{HMM EM Update}
\usage{
hmm2EMStep(K, y, pars, formula1, formula2, data)
}
\arguments{
\item{K}{the number of classes}

\item{y}{the sequence of observations}

\item{pars}{current sequence parameters}

\item{formula1}{formula for the logistic model relating the transition from state 1 to state 2 to the covariates}

\item{formula2}{formula for the logistic model relating the transition from state 2 to state 1 to the covariates}

\item{data}{a dataframe of covariates}
}
\value{
the updated set of parameters
}
\description{
The HMM update step of the EM algorithm
}
\details{
Perform a single step of the EM algorithm to fit the HMM component
of a two state multivariate Normal HMM model where the transistion
probabilities are governed by logistic regression models.
}

