% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MVNSeq.r
\name{initHMM2}
\alias{initHMM2}
\title{Initial Parameters}
\usage{
initHMM2(y, cl, formula1, formula2, data)
}
\arguments{
\item{y}{the sequence of observations}

\item{cl}{an integer vector allocating observations to classes}

\item{formula1}{formula for the logistic model relating the transition from state 1 to state 2 to the covariates}

\item{formula2}{formula for the logistic model relating the transition from state 2 to state 1 to the covariates}

\item{data}{a dataframe of covariates}
}
\value{
estimated model parameters
}
\description{
Initial parameter estimates for two state HMM model with variable
transition probabilities
}
\details{
Given a sequence of observations and a vector of initial class
allocations, estimate the parameters for a two state hidden Markov
model where the transition probabilities are governed by logistic
regression.
}

