% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MVNSeq.r
\name{initMix}
\alias{initMix}
\alias{initHMM}
\title{Initial Parameters}
\usage{
initMix(y, cl, K = max(cl))

initHMM(y, cl, K = max(cl))
}
\arguments{
\item{y}{the sequence of observations}

\item{cl}{an integer vector allocating observations to classes}

\item{K}{the number of classes}
}
\value{
`parsMixClass` returns a list with elements
}
\description{
Initial parameter estimates for HMM and mixture models.
}
\details{
Given a sequence of observations and a vector of initial class
allocations, estimate the parameters for a mixture or hidden
Markov model.
}
