% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MVNSeq.r
\name{mixEMStep}
\alias{mixEMStep}
\title{Mixture EM Update}
\usage{
mixEMStep(K, y, pars)
}
\arguments{
\item{K}{the number of classes}

\item{y}{the sequence of observations}

\item{pars}{current sequence parameters}
}
\value{
the updated set of parameters
}
\description{
The mixture update step of the EM algorithm.
}
\details{
Perform a single step of the EM algorithm to fit the mixture
component of a multivariate Normal mixture model.
}
