% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MVNSeq.r
\name{mvnEMStep}
\alias{gmvnEMStep}
\alias{grmvnEMStep}
\alias{mvnEMStep}
\title{MVN EM Update}
\usage{
mvnEMStep(K, y, pars)

gmvnEMStep(K, ys, pars)

grmvnEMStep(K, ys, pars, muv)
}
\arguments{
\item{K}{the number of classes}

\item{y}{the sequence of observations}

\item{pars}{current sequence parameters}

\item{ys}{the list of sequences of observations}

\item{muv}{current population parameters}
}
\value{
the updated set of parameters
}
\description{
The multivariate Normal update step of the EM algorithm.
}
\details{
Perform a single step of the EM algorithm
fitting a mean and covariance for each component.
}

