% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MVNSeq.r
\name{mvnHMM}
\alias{mvnHMM}
\alias{gmvnHMM}
\alias{grmvnHMM}
\title{Multivariate Normal Mixture Model}
\usage{
mvnHMM(
  y,
  cl,
  min.iters = 10,
  max.iters = 50,
  tol = 0.001,
  verbose = interactive()
)

gmvnHMM(
  y,
  cl,
  gr,
  common.transition = FALSE,
  min.iters = 10,
  max.iters = 50,
  tol = 0.001,
  verbose = interactive()
)

grmvnHMM(
  y,
  cl,
  gr,
  common.transition = FALSE,
  min.iters = 10,
  max.iters = 100,
  tol = 0.001,
  verbose = interactive()
)
}
\arguments{
\item{y}{the sequence of observations}

\item{cl}{an integer vector allocating observations to classes}

\item{min.iters}{minimum number of EM iterations}

\item{max.iters}{maximum number of EM iterations}

\item{tol}{tolerance for the log likelihood}

\item{verbose}{should the log likelihood be reported.}

\item{gr}{an integer vector allocating observations to groups}

\item{common.transition}{should the transition probabilities be
common across groups}
}
\value{
the fitted model
}
\description{
Fit multivariate Normal Hidden Markov models by EM.
}
\details{
These functions fit K state multivariate Normal hidden Markov
models to sequences of observations. `mvnHMM` fits a hidden Markov
model to a single sequence of observations. `gmvnHMM` and
`grmvnHMM` fits separate Markov models to several groups
(sequences) of observations. `gmvnHMM` fits a hidden Markov model
to each group so that each state has a different mean across
groups, but a common covariance. `grmvnHMM` constrains the means
of the states to be Normally distributed across groups.
}
