% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MVNSeq.r
\name{mvnHMM2}
\alias{mvnHMM2}
\alias{gmvnHMM2}
\alias{grmvnHMM2}
\title{Multivariate Normal Hidden Markov Model}
\usage{
mvnHMM2(
  y,
  cl,
  formula1,
  formula2,
  data,
  min.iters = 10,
  max.iters = 50,
  tol = 0.001,
  verbose = interactive()
)

gmvnHMM2(
  y,
  cl,
  gr,
  formula1,
  formula2,
  data,
  min.iters = 10,
  max.iters = 50,
  tol = 0.001,
  verbose = interactive()
)

grmvnHMM2(
  y,
  cl,
  gr,
  formula1,
  formula2,
  data,
  min.iters = 10,
  max.iters = 100,
  tol = 0.001,
  verbose = interactive()
)
}
\arguments{
\item{y}{the sequence of observations}

\item{cl}{an integer vector allocating observations to classes}

\item{formula1}{formula for the logistic model relating the
transition from state 1 to state 2 to the covariates}

\item{formula2}{formula for the logistic model relating the
transition from state 2 to state 1 to the covariates}

\item{data}{a dataframe of covariates}

\item{min.iters}{minimum number of EM iterations}

\item{max.iters}{maximum number of EM iterations}

\item{tol}{tolerance for the log likelihood}

\item{verbose}{should the log likelihood be reported.}

\item{gr}{an integer vector allocating observations to groups}
}
\value{
the fitted model
}
\description{
Fit multivariate Normal 2 State Hidden Markov models by EM.
}
\details{
These functions fit multivariate Normal 2 state hidden Markov
models to sequences of observations, where the transition
probabilities are governed by logistic regression. `mvnHMM2` fits
a hidden Markov model to a single sequence of observations.
`gmvnHMM2` and `grmvnHMM2` fit separate Markov models to several
groups (sequences) of observations. `gmvnHMM2` fits a hidden
Markov model to each group so that each state has a different mean
across groups, but a common covariance. `grmvnHMM2` constrains the
means of the states to be Normally distributed across groups.
}
