% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MVNSeq-package.R
\docType{package}
\name{MVNSeq-package}
\alias{MVNSeq}
\alias{MVNSeq-package}
\title{MVNSeq: Normal Mixture and Hidden Markov Models}
\description{
Fit Normal Mixture and Hidden Markov models to grouped data, with random effects to represent group variation.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/SWotherspoon/MVNSeq}
  \item Report bugs at \url{https://github.com/SWotherspoon/MVNSeq/issues}
}

}
\author{
\strong{Maintainer}: Simon Wotherspoon \email{Simon.Wotherspoon@utas.edu.au}

Authors:
\itemize{
  \item Simon Wotherspoon \email{Simon.Wotherspoon@utas.edu.au}
  \item Aidan Bindoff \email{Bindoffa@utas.edu.au}
}

}
\keyword{internal}
