% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MVNSeq.r
\name{initGroup}
\alias{initGroup}
\title{Initial Parameters}
\usage{
initGroup(K, pars)
}
\arguments{
\item{K}{the number of classes}

\item{pars}{lists of parameters generated by \code{initMix} or
\code{initHMM}}
}
\value{
the estimated parameters
}
\description{
Initial parameter estimates for random effect HMM and mixture models
}
\details{
Estimates the random effects parameters from lists of mixture or
hidden Markov parameters.
}
