% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MVNSeq.r
\docType{package}
\name{MVNSeq-package}
\alias{MVNSeq-package}
\title{Hidden Markov and Mixture models for sequences of multivariate
Normal observations.}
\description{
Provides facilities for fitting hidden Markov and mixture models to
sequences of multivariate Normal observations.  Allows for multiple
sequences and random variation in the state means across different
sequences.
}
\author{
S. Wotherspoon, A. Bindoff.
}
